requirements.txt
albumentations==0.5.2
filesplit==3.0.2
opencv-python==4.8.1.78
imgaug==0.4.0
kornia==0.5.0
pyclipper==1.3.0.post5
lmdb==1.4.1
PyYAML==6.0.1
omegaconf==2.1.2
tqdm==4.66.1
easydict==1.9
pandas==2.0.3
webdataset==0.2.57
pytorch-lightning==1.2.9
numpy==1.23.1
protobuf==3.20.0
av==11.0.0
einops==0.7.0